open-resource/requirements.txt
torch>=2.3
torchvision
pytorch-wavelets==1.3.0
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66
scikit-learn>=1.3
matplotlib>=3.8
